Split admin dashboard view shell
